[qca-nss-ecm] Add 64 bits support
- Define a new data type called ecm_ptr_t which is
  uint32_t for 32 bits systems and uint64_t for 64 bits
  systems.
- Declare the IPv6 create message instance as a pointer and
  allocate it from the heap. Since the size of the data types
  increased, the accelerate function is giving error for the
  frame size in 64 bits builds.

+++ ecm_state.c
@@ -691,7 +691,7 @@ static ssize_t ecm_state_char_device_read(struct file *file,	/* see include/linu
 
 	sfi = (struct ecm_state_file_instance *)file->private_data;
 	DEBUG_CHECK_MAGIC(sfi, ECM_STATE_FILE_INSTANCE_MAGIC, "%p: magic failed", sfi);
-	DEBUG_TRACE("%p: State read up to length %d bytes\n", sfi, length);
+	DEBUG_TRACE("%p: State read up to length %d bytes\n", sfi, (int)length);
 
 	/*
 	 * If there is still some message remaining to be output then complete that first
